[#36] Rename files related to parser
