+++ include/SCC/main.h
@@ -9,7 +9,7 @@
 #include "SCC/log/log.h"
 #include "SCC/query-assembler/query_assembler.h"
 #include "SCC/dump/tree_dump.h"
-#include "SCC/parser/syntax_analyzer.h"
+#include "SCC/parser/parser.h"
 
 #include "logger/initializers/console_appender_initializer.hpp"
 #include "logger/initializers/file_appender_initializer.hpp"
