+++ src/parser/CMakeLists.txt
@@ -1,10 +1,10 @@
 add_library(scc_parser STATIC)
 target_sources(scc_parser PRIVATE
-        syntax_analyzer_ddl.cpp
-        syntax_analyzer_dml.cpp
-        syntax_analyzer_general.cpp
-        syntax_analyzer_other.cpp
-        syntax_analyzer_utils.cpp
+        parser_ddl.cpp
+        parser_dml.cpp
+        parser_general.cpp
+        parser_other.cpp
+        utils.cpp
 )
 target_link_libraries(scc_parser PRIVATE
         scc_include
