+++ test/SQL_AST_test/SQL_AST_test.cpp
@@ -2,7 +2,7 @@
 #include "gmock/gmock.h"
 #include "SCC/ast/nodes/inode.h"
 #include "SCC/lexer/lexer.h"
-#include "SCC/parser/syntax_analyzer.h"
+#include "SCC/parser/parser.h"
 #include "SCC/dump/tree_dump.h"
 #include "logger/initializers/console_appender_initializer.hpp"
